pyproject.toml
[build-system]
requires = ["hatchling>=1.20.0"]
build-backend = "hatchling.build"

[project]
name = "context-packager"
version = "0.2.1"
description = "Convert documents to Markdown (MarkItDown), package with Repomix, and serve a Streamlit UI for copyable LLM context."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "Context Tools" }]
dependencies = [
  "repomix==0.3.4",
  "streamlit==1.50.0",
  "tiktoken==0.11.0",
  "pdfminer.six==20250506",
  "SQLAlchemy==2.0.35",
  "scikit-learn==1.5.2",
  "psycopg[binary,pool]==3.2.1",
  "pynndescent==0.5.13",
  "APScheduler==3.10.4",
]

[project.urls]
Homepage = "https://github.com/ruizmr/resume-context-builder"

[project.scripts]
context-ui = "resume_ui.cli:main"
resume-ui = "resume_ui.cli:main"
context-upsert = "resume_ui.upsert_cli:main"
context-ingest = "resume_ui.ingest_cli:main"
context-scheduler = "resume_ui.scheduler_cli:main"

[tool.hatch.build.targets.wheel]
packages = ["resume_ui", "hr_tools", "kb", "context_packager_data"]
force-include = { "hr_tools/instruction.md" = "hr_tools/instruction.md", "context_packager_data/tiktoken_cache" = "context_packager_data/tiktoken_cache" }

[tool.hatch.build.targets.sdist]
exclude = [
  "/dist",
  "/uploads",
  "/output",
  "/resume-dataset",
  "/resume-dataset-md",
  "/streamlit.log",
]

[project.optional-dependencies]
full = [
  "markitdown[all]==0.1.3",
  "onnxruntime==1.23.0; sys_platform == 'darwin'",
  "watchdog>=4.0.0",
]


